Referenced the JavaFX libraries and .jar MySQL connector file to implement a GUI and connect to the locally hosted MySQL hotel database.
Added the main class (App.java) which launches the app's GUI as well as the DBUtils.java class which is responsible for handling the opening and closing of MySQL database connections when performing database operations.

For security, MySQL database configuration information have been added to a .env file which is referenced in the .gitignore file to prevent it from mistakenly being committed to the Github repository in the future.