Created a User class (User.java) to instantiate a User object with the user's relevant details upon successful authentication by a hotel user.
In the created login functionality (UserLogin.java), a login scene is initialized and created upon launching the application in the main class (App.java) where the user is asked to input his/her username and password. The user's inputted password is compared to the stored password in the database and is authenticated if they match. A suitable error message is presented for invalid login, or failure to connect to the MySQL database.

For the hashing functionality in the Hashing class, I created the logic for appending randomly generated and secure salts per password generation and hashing them using a SHA-256 algorithm to prevent against rainbow table attacks. The reverse logic was applied for the password comparison operation upon login.

All SQL operations use prepared statements to prevent against any SQLi attacks resulting from malicious user input.