Added rate limiting functionality to user login mechanism, account unlock feature and sign out buttons added for Admin user, and renamed UserRegistration.java to AdminDashboard.java
I added the rate limiting security mechanism to the user login mechanism which locks a user's account after 5 failed login attempts. Once the account is locked, only the Admin user is able to manually unlock the account and allow access once again. Upon successful authentication, the failed attempts counter for the user's account is reset to 0. Two new columns have been added to the users table named "failedAttempts" and "isLocked".

The User Registration window has been rebranded to Admin Dashboard as it now includes not only the registration functionality, but the unlock account and sign out buttons as well.