Implemented validation for customer registration fields and QID value encryption.
Regex input validation was added to the customer registration fields (similar to what was done in the ValidateUserRegistration.java class).

AES-128 encryption algorithm was used with a 16-byte secret key to encrypt and decrypt a customer's QID value. The secret key is stored in the .env file. In order to store the encrypted QID value as a string in the database, the output is base64 encoded and is also base64 decoded as well in the decryption logic.

Also, the QID column in the database has been changed from BIGINT to VARCHAR to accomodate the encryption changes. Changes to the qid variable type have also been made in the source code.